Refactor API routers and add backend tests
