+++ backend/api/services/brevo.js
@@ -1,0 +1,84 @@
+const fetchImpl = (...args) => fetch(...args);
+
+function createBrevoService({ getSanityClient, logger, fetch = fetchImpl } = {}) {
+  const resolveHeaders = () => {
+    const key = process.env.BREVO_API_KEY;
+    if (!key) return null;
+    return {
+      'api-key': key,
+      accept: 'application/json',
+      'content-type': 'application/json',
+    };
+  };
+
+  const getHeaders = () => resolveHeaders();
+
+  const sendEmail = async (payload) => {
+    const headers = resolveHeaders();
+    if (!headers) {
+      const error = new Error('Email service not configured');
+      error.code = 'EMAIL_NOT_CONFIGURED';
+      throw error;
+    }
+    const response = await fetch('https://api.brevo.com/v3/smtp/email', {
+      method: 'POST',
+      headers,
+      body: JSON.stringify(payload),
+    });
+    if (!response.ok) {
+      const details = await response.text().catch(() => '');
+      const error = new Error('Brevo send failed');
+      error.status = response.status;
+      error.details = details;
+      throw error;
+    }
+    return response;
+  };
+
+  const upsertContact = async ({ email, firstName, lastName, phone }) => {
+    const headers = resolveHeaders();
+    if (!headers) throw new Error('BREVO_API_KEY is not configured on the server');
+    const body = {
+      email,
+      attributes: {
+        FIRSTNAME: firstName || undefined,
+        LASTNAME: lastName || undefined,
+        SMS: phone || undefined,
+      },
+      updateEnabled: true,
+    };
+    const response = await fetch('https://api.brevo.com/v3/contacts', {
+      method: 'POST',
+      headers,
+      body: JSON.stringify(body),
+    });
+    if (!response.ok && response.status !== 400) {
+      const text = await response.text().catch(() => '');
+      throw new Error(`Brevo contacts error ${response.status}: ${text}`);
+    }
+    const sc = getSanityClient ? getSanityClient() : null;
+    if (sc && email) {
+      try {
+        await sc.createIfNotExists({
+          _id: `contact-${email}`,
+          _type: 'contact',
+          email,
+          firstName: firstName || null,
+          lastName: lastName || null,
+          phone: phone || null,
+          updatedAt: new Date().toISOString(),
+        });
+      } catch (err) {
+        if (logger) logger.warn({ err }, 'failed to mirror contact in sanity');
+      }
+    }
+  };
+
+  return {
+    getHeaders,
+    sendEmail,
+    upsertContact,
+  };
+}
+
+module.exports = { createBrevoService };
