+++ package.json
@@ -77,10 +77,12 @@
     "lint-staged": "^16.1.5",
     "pino-pretty": "^11.2.2",
     "prettier": "^3.6.2",
+    "supertest": "^6.3.4",
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
     "typescript": "~4.7.4",
     "vite": "^5.1.7",
+    "vitest": "^1.6.1",
     "webpack-bundle-analyzer": "^4.10.2"
   },
   "husky": {
